+++ opyoid/bindings/__init__.py
@@ -10,4 +10,4 @@ from .multi_binding import ItemBinding, ListProvider, MultiBinding, MultiBinding
 from .private_module import PrivateModule
 from .provider_binding import FromProviderProvider, ProviderBinding, ProviderBindingToProviderAdapter
 from .registered_binding import RegisteredBinding
-from .self_binding import FromClassProvider, SelfBinding, SelfBindingToProviderAdapter
+from .self_binding import FromCallableProvider, SelfBinding, SelfBindingToProviderAdapter