+++ opyoid/bindings/provider_binding/provider_binding_to_provider_adapter.py
@@ -8,19 +8,27 @@ from opyoid.target import Target
 from opyoid.utils import InjectedT
 from .from_provider_provider import FromProviderProvider
 from .provider_binding import ProviderBinding
+from ..self_binding import CallableToProviderAdapter
 
 
 class ProviderBindingToProviderAdapter(BindingToProviderAdapter[ProviderBinding]):
     """Creates a Provider from a ProviderBinding."""
 
+    def __init__(self):
+        BindingToProviderAdapter.__init__(self)
+        self._adapter = CallableToProviderAdapter()
+
     def accept(self, binding: Binding[InjectedT], context: InjectionContext) -> bool:
         return isinstance(binding, ProviderBinding)
 
-    def create(self,
-               binding: RegisteredBinding[ProviderBinding[InjectedT]],
-               context: InjectionContext[InjectedT]) -> Provider[InjectedT]:
+    def create(
+        self, binding: RegisteredBinding[ProviderBinding[InjectedT]], context: InjectionContext[InjectedT]
+    ) -> Provider[InjectedT]:
         if isinstance(binding.raw_binding.bound_provider, Provider):
             return binding.raw_binding.bound_provider
+        if not isinstance(binding.raw_binding.bound_provider, type):
+            return self._adapter.create(binding, binding.raw_binding.bound_provider, context)
+        # noinspection PyTypeChecker
         provider_target = Target(binding.raw_binding.bound_provider, binding.raw_binding.named)
         provider_context = context.get_child_context(provider_target)
         provider_provider = provider_context.get_provider()
@@ -31,6 +39,8 @@ class ProviderBindingToProviderAdapter(BindingToProviderAdapter[ProviderBinding]
         try:
             scope_provider = scope_context.get_provider()
         except NoBindingFound:
-            raise NonInjectableTypeError(f"Could not create a provider for {binding}: they are no bindings for"
-                                         f"the scope {binding.raw_binding.scope}") from None
+            raise NonInjectableTypeError(
+                f"Could not create a provider for {binding}: they are no bindings for"
+                f"the scope {binding.raw_binding.scope}"
+            ) from None
         return scope_provider.get().get_scoped_provider(unscoped_provider)