+++ opyoid/providers/providers_factories/tuple_provider_factory.py
@@ -1,6 +1,6 @@
 from typing import List, Tuple
 
-from opyoid.bindings import FromClassProvider
+from opyoid.bindings import FromCallableProvider
 from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.target import Target
@@ -18,4 +18,4 @@ class TupleProviderFactory(ProviderFactory):
     def create(self, context: InjectionContext[Tuple[InjectedT]]) -> Provider[Tuple[InjectedT]]:
         new_target = Target(List[context.target.type.__args__[0]], context.target.named)
         new_context = context.get_child_context(new_target)
-        return FromClassProvider(tuple, [new_context.get_provider()], None, {})
+        return FromCallableProvider(tuple, [new_context.get_provider()], None, {})