+++ tests/test_providers/test_provider_creator.py
@@ -3,8 +3,16 @@ from typing import List, Optional, Set, Tuple, Type
 from unittest.mock import ANY
 
 from opyoid import Provider, SelfBinding
-from opyoid.bindings import BindingRegistry, ClassBinding, FromClassProvider, FromInstanceProvider, \
-    InstanceBinding, ListProvider, MultiBinding, ProviderBinding
+from opyoid.bindings import (
+    BindingRegistry,
+    ClassBinding,
+    FromCallableProvider,
+    FromInstanceProvider,
+    InstanceBinding,
+    ListProvider,
+    MultiBinding,
+    ProviderBinding,
+)
 from opyoid.bindings.multi_binding import ItemBinding
 from opyoid.bindings.registered_binding import RegisteredBinding
 from opyoid.bindings.registered_multi_binding import RegisteredMultiBinding
@@ -163,7 +171,7 @@ class TestProviderCreator(unittest.TestCase):
         )
         context = InjectionContext(Target(Set[MyType]), self.state)
         provider = self.provider_creator.get_provider(context)
-        self.assertIsInstance(provider, FromClassProvider)
+        self.assertIsInstance(provider, FromCallableProvider)
         set_instance = provider.get()
         self.assertIn(self.my_instance, set_instance)
         self.assertEqual(2, len(set_instance))
@@ -186,7 +194,7 @@ class TestProviderCreator(unittest.TestCase):
         )
         context = InjectionContext(Target(Tuple[MyType]), self.state)
         provider = self.provider_creator.get_provider(context)
-        self.assertIsInstance(provider, FromClassProvider)
+        self.assertIsInstance(provider, FromCallableProvider)
         tuple_instance = provider.get()
         self.assertEqual((self.my_instance, ANY), tuple_instance)
         self.assertIsInstance(tuple_instance[1], MyType)